Create The Skyline Problem
The skyline should be represented as a list of "key points" sorted by their x-coordinate in the form [[x1,y1],[x2,y2],...]. Each key point is the left endpoint of some horizontal segment in the skyline except the last point in the list, which always has a y-coordinate 0 and is used to mark the skyline's termination where the rightmost building ends. Any ground between the leftmost and rightmost buildings should be part of the skyline's contour.
some comments to understand the code
"""
type buildings: List[List[int]]
rtype: List[List[int]]
        """
        # `position` stores all coordinates where the largest height may change
        # `alive` stores all buildings whose ranges cover the current coordinate
         # pop buildings that end at or before `curPos` out of the priority queue
            # they are no longer "alive"
          # push [negative_height, end_point] of all buildings that start before `curPos` onto the priority queue
            # they are candidates for the current highest building
          # now alive[0] must be the largest height at the current position

Brilliant! verbosely commented code to make it even easier to understand.
It's also worth noting that when we remove buildings that are no longer alive from the priority queue, there is no need to remove all "dead" buildings, as long as they are low enough and wouldn't stop us from finding the highest alive building. For instance, if the input is [ [1, 4, 20], [2, 3, 10] ], when we reach x = 3, the building of height 10 is still in the queue, but it doesn't matter.
I mention this because I've tried to turn this answer into Python code but got time exceeded. That code tries to remove an element (not the smallest one) from the priority queue, which is an O(N) operation. 
another method  or approach can be- 
A Simple Solution is to initialize skyline or result as empty, then one by one add buildings to skyline. A building is added by first finding the overlapping strip(s). If there are no overlapping strips, the new building adds new strip(s). If overlapping strip is found, then height of the existing strip may increase. Time complexity of this solution is O(n2)