Create Palindrome.cpp
A palindrome is a word, number, phrase, or other sequence of symbols that reads the same backwards as forwards, such as the words madam or racecar, the date/time stamps 11/11/11 11:11 and 02/02/2020, and the sentence: "A man, a plan, a canal – Panama". The 19-letter Finnish word saippuakivikauppias (a soapstone vendor), is the longest single-word palindrome in everyday use, while the 12-letter term tattarrattat (from James Joyce in Ulysses) is the longest in English.

The word palindrome was introduced by English poet and writer Henry Peacham in 1638.[1] The concept of a palindrome can be dated to the 3rd-century BCE, although no examples survive; the first physical examples can be dated to the 1st-century CE with the Latin acrostic word square, the Sator Square (contains both word and sentence palindromes), and the 4th-century Greek Byzantine sentence palindrome nipson anomemata me monan opsin.[2][3]

Palindrome are also found in music (the table canon and crab canon) and biological structures (most genomes include palindromic gene sequences). In automata theory, the set of all palindromes over an alphabet is a language that is context-free, but not regular.